cannot work with t_eat

+++ utils.c
@@ -6,7 +6,7 @@
 /*   By: msalena <msalena@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/09 17:55:13 by msalena           #+#    #+#             */
-/*   Updated: 2022/01/15 18:53:18 by msalena          ###   ########.fr       */
+/*   Updated: 2022/01/16 13:10:59 by msalena          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,13 +14,13 @@
 
 /*	Time which was spending from the starting
 	of threads time to actual time */
-long	actual_time(struct timeval *start_time)
+unsigned long	actual_time(struct timeval *start_time)
 {
 	struct timeval	cur_time;
 
 	gettimeofday(&cur_time, NULL);
-	return ((long)((cur_time.tv_usec - start_time->tv_usec) / 1000)
-			+ ((cur_time.tv_sec - start_time->tv_sec) * 1000));
+	return ((unsigned long)((cur_time.tv_usec - start_time->tv_usec) / 1000)
+			+ (unsigned long)((cur_time.tv_sec - start_time->tv_sec) * 1000));
 }
 
 long	tmp_micsec_AT(struct timeval *start_time)
@@ -28,8 +28,8 @@ long	tmp_micsec_AT(struct timeval *start_time)
 	struct timeval	cur_time;
 
 	gettimeofday(&cur_time, NULL);
-	return ((long)(cur_time.tv_usec - start_time->tv_usec)
-			+ ((cur_time.tv_sec - start_time->tv_sec) * 1000000));
+	return ((unsigned long)(cur_time.tv_usec - start_time->tv_usec)
+			+ (unsigned long)((cur_time.tv_sec - start_time->tv_sec) * 1000000));
 }
 
 void	ft_usleep(int mlsec)
@@ -39,7 +39,7 @@ void	ft_usleep(int mlsec)
 
 	gettimeofday(&cur_time, NULL);
 	gettimeofday(&star_platinum, NULL);
-	while ((actual_time(&star_platinum) - actual_time(&cur_time)) < mlsec)
+	while ((actual_time(&star_platinum) - actual_time(&cur_time)) < (unsigned long)mlsec)
 	{
 		usleep (100);
 		gettimeofday(&cur_time, NULL);
